chatbot/requirements.txt
# Web/API layer
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# LLM / Reasoning / Graph
openai>=1.40.0
langchain>=0.3.27
langchain-core>=0.3.80
langchain-openai>=0.3.35
langgraph

# Environment / Utils
python-dotenv>=1.0.1
pydantic

# Vector DB
pinecone-client
langchain-pinecone

# PDF Processing
PyMuPDF==1.23.25

# Text Splitting
langchain-text-splitters

# 기본 데이터 처리
numpy>=1.26.2
pandas>=2.1.4
requests>=2.31.0
scikit-learn>=1.3.2
scipy>=1.11.4

# 이미지 처리
pillow>=10.0.0
python-multipart>=0.0.6

# Transformers (CLIP)
transformers>=4.40.0